Render a badge on system preview images https://bugs.webkit.org/show_bug.cgi?id=184854 <rdar://problem/39615154>
Reviewed by Tim Horton.

Source/WebCore:

At paint time, if a RenderImage is an image element
which is identified as a system preview, then draw
a little badge in the top right corner.

It is expected that platforms will override the
default rendering with something that identifies
how a system preview will operate. e.g. QuickLook
on Apple systems.

Test: system-preview/badge.html

* rendering/RenderImage.cpp:
(WebCore::RenderImage::paintIntoRect): If we are
a system preview, call RenderTheme to draw a badge.
* rendering/RenderTheme.cpp:
(WebCore::RenderTheme::paintSystemPreviewBadge): A default
implementation that draws a red circle.
* rendering/RenderTheme.h:

LayoutTests:

* system-preview/badge-expected.html: Added.
* system-preview/badge.html: Added.

Identifier: 200348@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230872 268f45cc-cd09-0410-ab3c-d52691b4dbfc