REGRESSION(r230950): Faulty commit sequencing in CoordinatedGraphicsScene https://bugs.webkit.org/show_bug.cgi?id=184917
Reviewed by Michael Catanzaro.

After r230950, current animation state for a given layer is also taken
into account when determining whether or not the layer requires a
backing store. For that to work properly, all the animation state has
to be updated before the backing store work. This patch changes the
order of helper method invocations in
CoordinatedGraphicsScene::setLayerState() to address that.

* Shared/CoordinatedGraphics/CoordinatedGraphicsScene.cpp:
(WebKit::CoordinatedGraphicsScene::setLayerState):


Identifier: 200428@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230954 268f45cc-cd09-0410-ab3c-d52691b4dbfc